requirements.txt
clip==0.2.0
clip_anytorch==2.5.0
numpy==1.21.6
opencv_python==4.6.0.66
pandas==1.3.5
Pillow==9.4.0
requests==2.28.1
scikit_learn==1.2.1
tensorboardX==2.6
torch==1.9.1+cu111
torchvision==0.10.1+cu111
tqdm==4.64.1
transformers==4.24.0
